create isolate component stacked-slides for template review app
